safari-600.7-branch/LayoutTests/platform/mac/fast/canvas/canvas-scale-strokePath-shadow-expected.txt

Ensure correct behavior of canvas with path stroke + shadow after scaling. A blue and red checkered pattern should be displayed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 76
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 200 (actual: 170)
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 49
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 199 (actual: 167)
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 70
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 70
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 69
PASS successfullyParsed is true

TEST COMPLETE
